Move failing test to WPCom from Jetpack
