+++ example/src/androidTest/java/org/wordpress/android/fluxc/release/ReleaseStack_AppComponent.java
@@ -31,6 +31,7 @@ public interface ReleaseStack_AppComponent {
     void inject(ReleaseStack_ActivityLogTestJetpack test);
     void inject(ReleaseStack_InsightsTestJetpack test);
     void inject(ReleaseStack_TimeStatsTestJetpack test);
+    void inject(ReleaseStack_TimeStatsTestWPCom test);
     void inject(ReleaseStack_ManageInsightsTestJetpack test);
     void inject(ReleaseStack_CommentTestWPCom test);
     void inject(ReleaseStack_CommentTestXMLRPC test);
